+++ GameOfLife/checkbuttonpressed.h
@@ -15,7 +15,7 @@ public:
     CheckButtonPressed();
     void checkStatusOfButtonPressed();
     bool getStatusOfLoop();
-    int getHowManySec();
+    float getHowManySec();
     bool getSaveToFileStatus();
     void setSaveToFileStatus(bool value);
 };
